clipping nigger

+++ lucid/lucid/renderer/vector/vector2d.h
@@ -1,10 +1,10 @@
 #pragma once
 
 struct vec2_t {
-    float x, y;
+    int x, y;
 
     vec2_t() {}
-    vec2_t(float _x, float _y)
+    vec2_t(int _x, int _y)
         : x(_x), y(_y) {}
 
     vec2_t operator + (const vec2_t& other) {
